refactor : change common package structure as context
