+++ src/main/java/coz/weavon/common/domain/model/Property.java
@@ -1,4 +1,4 @@
-package coz.weavon.common.schema;
+package coz.weavon.common.domain.model;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
