+++ src/main/java/coz/weavon/common/presentation/model/reqeuest/PageRequest.java
@@ -1,4 +1,4 @@
-package coz.weavon.common.io.model.reqeuest;
+package coz.weavon.common.presentation.model.reqeuest;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
