+++ src/main/java/coz/weavon/common/presentation/model/response/ResponseType.java
@@ -1,4 +1,4 @@
-package coz.weavon.common.io.model.response;
+package coz.weavon.common.presentation.model.response;
 
 public enum ResponseType {
     SUCCESS,
